Remove non-functional wrapper.
